requirement.txt
django
pymavlink
pyserial 